Mirror main

+++ api/src/controllers/users.js
@@ -1,4 +1,4 @@
-const {Address,User} = require('../db');
+const {User} = require('../db');
 
 async function createUser (name, lastName, dni, email, celphone, username, password, isAdmin, userAddress) {
     if (!(name && lastName && dni && email && celphone && username && password))
@@ -6,7 +6,7 @@ async function createUser (name, lastName, dni, email, celphone, username, passw
 
     const user = await User.create({name, lastName, dni, email, celphone, username, password, isAdmin});
     await user.setAddress(userAddress);
-    return user;
+    return await user.save();
 }
 
 async function getAdminUsers() {
@@ -36,17 +36,22 @@ async function getUserByUsername(username) {
         where,
         attributes
     });
+    if (!user) throw new Error('El usuario no existe en la base de datos.');
     return user;
 }
 
 async function updateUser(username,userData) {
     const where = {username};
-    let user = await User.update(userData, {where});
-    return user;
+    let user =  await User.findOne({where});
+    if (!user) throw new Error('El usuario no existe en la base de datos.');
+    user.set(userData);
+    return await user.save();
 }
 
 async function deleteUser(username) {
     const where = {username};
+    let user =  await User.findOne({where});
+    if (!user) throw new Error('El usuario no existe en la base de datos.');
     return await User.destroy({where});
 }
